# openapi.yml
openapi: 3.0.1
info:
    title: Secretariat
    version: 0.0.1
paths:
...
    /{pid}/{key}/latest:
        parameters:
        -   $ref: '#/components/parameters/pid'
        -   $ref: '#/components/parameters/key'
        get:
            summary: Get the latest value
            operationId: getLatest
            responses:
                200:
                    description: Successful
                    content: 
                        application/json: 
                            schema: 
                                $ref: '#/components/schemas/ValueAndTime'
                404:
                    description: Not found
components:
    schemas:
...
        JsonValue:
            nullable: true
            anyOf:
            - type: string
            - type: number
            - type: integer
            - type: boolean
            - type: array
              items: {}
            - type: object
        ValueAndTime:
            type: object
            properties:
                value: 
                    $ref: '#/components/schemas/JsonValue'
                time: 
                    type: number
    parameters:
        pid:
            in: path
            name: pid
            description: Project ID
            required: true
            schema: 
                type: string
                format: kebab-case
                minLength: 1
                maxLength: 32
        key:
            in: path
            name: key
            description: Name of the posted value
            required: true
            schema:
                type: string
                format: kebab-case
                minLength: 1
                maxLength: 16